+++ packit_service/worker/events/github/pr.py
@@ -1,0 +1,94 @@
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
+
+from typing import Optional
+
+from ogr.abstract import Comment as OgrComment
+from ogr.abstract import GitProject
+
+from packit_service.service.db_project_events import AddPullRequestEventToDb
+from packit_service.worker.events.comment import AbstractPRCommentEvent
+from packit_service.worker.events.enums import (
+    PullRequestAction,
+    PullRequestCommentAction,
+)
+from packit_service.worker.events.github.abstract import GithubEvent
+
+
+class Synchronize(AddPullRequestEventToDb, GithubEvent):
+    def __init__(
+        self,
+        action: PullRequestAction,
+        pr_id: int,
+        base_repo_namespace: str,
+        base_repo_name: str,
+        base_ref: str,
+        target_repo_namespace: str,
+        target_repo_name: str,
+        project_url: str,
+        commit_sha: str,
+        actor: str,
+    ) -> None:
+        super().__init__(project_url=project_url, pr_id=pr_id)
+        self.action = action
+        self.base_repo_namespace = base_repo_namespace
+        self.base_repo_name = base_repo_name
+        self.base_ref = base_ref
+        self.target_repo_namespace = target_repo_namespace
+        self.target_repo_name = target_repo_name
+        self.commit_sha = commit_sha
+        self.actor = actor
+        self.identifier = str(pr_id)
+        self.git_ref = None  # pr_id will be used for checkout
+
+    def get_dict(self, default_dict: Optional[dict] = None) -> dict:
+        result = super().get_dict()
+        result["action"] = result["action"].value
+        return result
+
+    def get_base_project(self) -> Optional[GitProject]:
+        return None  # With Github app, we cannot work with fork repo
+
+
+class Comment(AbstractPRCommentEvent, GithubEvent):
+    def __init__(
+        self,
+        action: PullRequestCommentAction,
+        pr_id: int,
+        base_repo_namespace: str,
+        base_repo_name: Optional[str],
+        base_ref: Optional[str],
+        target_repo_namespace: str,
+        target_repo_name: str,
+        project_url: str,
+        actor: str,
+        comment: str,
+        comment_id: int,
+        commit_sha: Optional[str] = None,
+        comment_object: Optional[OgrComment] = None,
+    ) -> None:
+        super().__init__(
+            pr_id=pr_id,
+            project_url=project_url,
+            comment=comment,
+            comment_id=comment_id,
+            commit_sha=commit_sha,
+            comment_object=comment_object,
+        )
+        self.action = action
+        self.base_repo_namespace = base_repo_namespace
+        self.base_repo_name = base_repo_name
+        self.base_ref = base_ref
+        self.target_repo_namespace = target_repo_namespace
+        self.target_repo_name = target_repo_name
+        self.actor = actor
+        self.identifier = str(pr_id)
+        self.git_ref = None  # pr_id will be used for checkout
+
+    def get_dict(self, default_dict: Optional[dict] = None) -> dict:
+        result = super().get_dict()
+        result["action"] = result["action"].value
+        return result
+
+    def get_base_project(self) -> Optional[GitProject]:
+        return None  # With Github app, we cannot work with fork repo
