+++ packit_service/worker/handlers/forges.py
@@ -26,8 +26,8 @@ from packit_service.worker.allowlist import Allowlist
 from packit_service.worker.checker.abstract import Checker
 from packit_service.worker.checker.forges import IsIssueInNotificationRepoChecker
 from packit_service.worker.events import (
-    InstallationEvent,
     IssueCommentEvent,
+    github,
 )
 from packit_service.worker.handlers.abstract import (
     JobHandler,
@@ -44,7 +44,7 @@ from packit_service.worker.result import TaskResults
 logger = logging.getLogger(__name__)
 
 
-@reacts_to(event=InstallationEvent)
+@reacts_to(event=github.installation.Installation)
 class GithubAppInstallationHandler(
     JobHandler,
     ConfigFromEventMixin,
@@ -66,7 +66,7 @@ class GithubAppInstallationHandler(
             event=event,
         )
 
-        self.installation_event = InstallationEvent.from_event_dict(event)
+        self.installation_event = github.installation.Installation.from_event_dict(event)
         self.account_type = self.installation_event.account_type
         self.account_login = self.installation_event.account_login
         self.sender_login = self.installation_event.sender_login
