+++ packit_service/worker/handlers/mixin.py
@@ -23,7 +23,9 @@ from packit_service.models import (
 from packit_service.utils import get_packit_commands_from_comment
 from packit_service.worker.events.copr import AbstractCoprBuildEvent
 from packit_service.worker.events.event import EventData
-from packit_service.worker.events.github import PullRequestCommentGithubEvent
+from packit_service.worker.events.github.pr import (
+    Comment as PullRequestCommentGithubEvent,
+)
 from packit_service.worker.events.gitlab import MergeRequestCommentGitlabEvent
 from packit_service.worker.events.koji import KojiBuildEvent, KojiBuildTagEvent
 from packit_service.worker.events.pagure import PullRequestCommentPagureEvent
