+++ packit_service/worker/jobs.py
@@ -32,10 +32,10 @@ from packit_service.worker.events import (
     CheckRerunEvent,
     Event,
     EventData,
-    InstallationEvent,
     IssueCommentEvent,
     PullRequestCommentPagureEvent,
     PullRequestPagureEvent,
+    github,
 )
 from packit_service.worker.events.comment import (
     AbstractCommentEvent,
@@ -221,7 +221,7 @@ class SteveJobs:
 
         # installation is handled differently b/c app is installed to GitHub account
         # not repository, so package config with jobs is missing
-        if isinstance(self.event, InstallationEvent):
+        if isinstance(self.event, github.installation.Installation):
             GithubAppInstallationHandler.get_signature(
                 event=self.event,
                 job=None,
