+++ tests/unit/test_srpm_logs.py
@@ -19,12 +19,14 @@ from packit.config import (
 
 from packit_service.config import ServiceConfig
 from packit_service.models import SRPMBuildModel
-from packit_service.worker.events.github import (
-    PullRequestCommentGithubEvent,
-    PullRequestGithubEvent,
-    PushGitHubEvent,
-    ReleaseEvent,
+from packit_service.worker.events.github.pr import (
+    Comment as PullRequestCommentGithubEvent,
 )
+from packit_service.worker.events.github.pr import (
+    Synchronize as PullRequestGithubEvent,
+)
+from packit_service.worker.events.github.push import Push as PushGitHubEvent
+from packit_service.worker.events.github.release import Release as ReleaseEvent
 from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
 
 logger = logging.getLogger(__name__)
